+++ tekore/_sender/__init__.py
@@ -2,3 +2,16 @@ from .base import Sender, Request, Response
 from .concrete import SyncSender, AsyncSender
 from .extending import ExtendingSender, RetryingSender, CachingSender
 from .client import SenderConflictWarning, Client, send_and_process
+from .error import (
+    HTTPError,
+    ClientError,
+    ServerError,
+    BadRequest,
+    Unauthorised,
+    Forbidden,
+    NotFound,
+    TooManyRequests,
+    InternalServerError,
+    BadGateway,
+    ServiceUnavailable,
+)